Added type validation for email and number
Had to do a quick bit of research on this as it
was letting me input words into the tel
box even after I added the tel type. Turns
out that adding a pattern fixes this issues.

Will likely be a while before I learn all
the regex for the patterns but at least
I will likely know that this will solve the
problem and I can look it up going forward.